feat(settings): split settings into categorized full-screen overlay
Replace the single-scroll settings tab in the workspace panel with a
Proma-style full-screen settings overlay: left sidebar category nav +
right content area, one component file per category.

- New react/settings/ directory: SettingsOverlay shell plus General
  (theme/language), Models (list/edit/new provider), Memory (Simple
  Mode + L1/L2/L3), Integrations (content hub/OCR/Tavily), Channels
  (Feishu/WeChat) and About (health/wiki stats) pages, with shared
  SettingsSection/SettingsCard/SettingsRow primitives.
- app-store gains settingsOpen/activeSettingsTab state; ?tab=settings
  deep link still opens the overlay on boot.
- WorkspacePanel drops the settings tab in favor of a gear button that
  opens the overlay; Simple Mode hides integrations/channels nav items
  and locks memory toggles as before.
