+++ apps/inno-agent/web/src/react/settings/shared.ts
@@ -1,0 +1,15 @@
+import type { InnoModelInfo } from "../../types/settings.js";
+
+export function formatBytes(bytes: number): string {
+	if (bytes < 1024) return `${bytes} B`;
+	if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
+	return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
+}
+
+export function formatTokens(tokens: number): string {
+	return tokens >= 1000 ? `${Math.round(tokens / 1000)}k` : String(tokens);
+}
+
+export function modelKey(model: InnoModelInfo): string {
+	return `${model.provider}:${model.id}`;
+}
